# server/api/openapi.yaml
openapi: 3.0.3
info:
  title: AgentTeams API
  description: Enterprise Windows Machine Management System API
  version: 1.0.0
  contact:
    name: AgentTeams Team
servers:
  - url: https://api.agentteams.io/api/v1
    description: Production server
...
paths:
...
  /tasks:
    get:
      tags: [Tasks]
      summary: List tasks
      security:
        - bearerAuth: []
      parameters:
        - name: agent_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, success, failed, timeout, cancelled]
        - name: page
          in: query
          schema:
            type: integer
        - name: page_size
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
    post:
      tags: [Tasks]
      summary: Create task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
  /tasks/batch:
    post:
      tags: [Tasks]
      summary: Create multiple tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchTaskRequest'
      responses:
        '201':
          description: Tasks created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchTaskResponse'
  /tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get task details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
    delete:
      tags: [Tasks]
      summary: Cancel task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task cancelled
  /tasks/{id}/output:
    get:
      tags: [Tasks]
      summary: Stream task output (SSE)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task output stream
          content:
            text/event-stream:
              schema:
                type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateTaskRequest:
      type: object
      required:
        - agent_id
        - command_type
      properties:
        agent_id:
          type: string
        command_type:
          type: string
          enum: [exec_shell, init_machine, clean_disk]
        params:
          type: object
        timeout:
          type: integer
          default: 300
    BatchTaskRequest:
      type: object
      properties:
        agent_ids:
          type: array
          items:
            type: string
        command_type:
          type: string
        params:
          type: object
        timeout:
          type: integer
    TaskResponse:
      type: object
      properties:
        id:
          type: string
        agent_id:
          type: string
        command_type:
          type: string
        status:
          type: string
          enum: [pending, running, success, failed, timeout, cancelled]
        exit_code:
          type: integer
        output:
          type: string
        duration:
          type: number
        created_at:
          type: string
          format: date-time
    TaskListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer